Small chanages for pull request
