+++ app/src/main/java/com/ftn/trippleaproject/repository/local/database/dao/room/EventDao.java
@@ -26,5 +26,5 @@ public interface EventDao {
     Flowable<List<EventDb>> readAll();
 
     @Delete
-    void delete(EventDb eventDb);
+    void delete(EventDb... eventDb);
 }
